EUC-JISX0213の文字列の読み込みをDataから行うようにする
